create side bar
I have made the new project and created the "index.html" and the "styles.css" files. I have also added the "fonts" and "SVG's" folder. The fonts folder contains roboto fonts and the svg folders currently contains the icons used for the sidebar. So far the sidebar is mostly complete now.