add bottom half of the top portion
The top portion of the page is now completed. I have added indicators in the "top-portion-styles.css" file to show which css is for the top half and which is for the bottom half. I have also fixed the padding and margins for the top portion so both halfs line up correctly. In the "sidebar-styles.css" file i have changed the spacing between the buttons on the sidebar as well.