create annoucements portion of the page
Edited the "index.html" file to add the "<div id="announcements">" div to the page that contain all the annoucements html. Edited the "annoucements-styles.css" file to styles the annoucements part of the page. Edited line 39-42 in "project-styles.css" to add a bigger gap between the project title and project description. Edited "index.html" to include trending title. Edited "trending-styles.css" to include styling for trending title. Edited line 10 in "main-content-styles.css" to include a row gap to make a gap between the announcements and trending.